+++ changelog.md
@@ -1,4 +1,4 @@
- - Convert chains between evenly spaced and unspaced chains from the slice panel. Works with DC, Octatrack, and OP-x slice data.
+ - Convert chains between evenly spaced and unspaced chains from the slice panel. Works with DC, Octatrack, and OP-x slice data. When converted to a spaced chain, if the grid-size is set to a value, the new chain will be padded to this number.
  - Ctrl+Click to preview slices before slicing a sample.
 
 2023-07-27 1.4.5